requirements.txt
streamlit==1.54.0
pandas==2.2.0
openpyxl==3.1.5
Pillow==10.4.0
streamlit-image-coordinates==0.4.0
